+++ starlite/utils/url.py
@@ -1,3 +1,4 @@
+import re
 from typing import Sequence
 
 
@@ -5,14 +6,10 @@ def normalize_path(path: str) -> str:
     """
     Normalizes a given path by ensuring it starts with a slash and does not end with a slash
     """
-    if path == "/":
-        return path
+    path = path.rstrip("/")
     if not path.startswith("/"):
         path = "/" + path
-    if path.endswith("/"):
-        path = path[: len(path) - 1]
-    while "//" in path:
-        path = path.replace("//", "/")
+    path = re.sub("//+", "/", path)
     return path
 
 
@@ -20,8 +17,4 @@ def join_paths(paths: Sequence[str]) -> str:
     """
     Normalizes and joins path fragments
     """
-    normalized_fragments = []
-    for fragment in paths:
-        fragment = normalize_path(fragment)
-        normalized_fragments.append(fragment)
-    return "".join(normalized_fragments)
+    return normalize_path("/".join(paths))