+++ tests/openapi/test_tags.py
@@ -1,12 +1,14 @@
-from typing import Any, Type, cast
+from typing import TYPE_CHECKING, Any, Type, cast
 
 import pytest
-from openapi_schema_pydantic.v3.v3_1_0.open_api import OpenAPI
 
 from starlite import Controller, HTTPRouteHandler, Router, Starlite, get
 
+if TYPE_CHECKING:
+    from pydantic_openapi_schema.v3_1_0.open_api import OpenAPI
 
-@pytest.fixture
+
+@pytest.fixture()
 def handler() -> HTTPRouteHandler:
     @get("/handler", tags=["handler"])
     def _handler() -> Any:
@@ -15,7 +17,7 @@ def handler() -> HTTPRouteHandler:
     return _handler
 
 
-@pytest.fixture
+@pytest.fixture()
 def controller() -> Type[Controller]:
     class _Controller(Controller):
         path = "/controller"
@@ -28,28 +30,28 @@ def controller() -> Type[Controller]:
     return _Controller
 
 
-@pytest.fixture
+@pytest.fixture()
 def router(controller: Type[Controller]) -> Router:
     return Router(path="/router", route_handlers=[controller], tags=["router"])
 
 
-@pytest.fixture
+@pytest.fixture()
 def app(handler: HTTPRouteHandler, controller: Type[Controller], router: Router) -> Starlite:
     return Starlite(route_handlers=[handler, controller, router])
 
 
-@pytest.fixture
-def openapi_schema(app: Starlite) -> OpenAPI:
-    return cast(OpenAPI, app.openapi_schema)
+@pytest.fixture()
+def openapi_schema(app: Starlite) -> "OpenAPI":
+    return cast("OpenAPI", app.openapi_schema)
 
 
-def test_openapi_schema_handler_tags(openapi_schema: OpenAPI) -> None:
+def test_openapi_schema_handler_tags(openapi_schema: "OpenAPI") -> None:
     assert openapi_schema.paths["/handler"].get.tags == ["handler"]  # type: ignore
 
 
-def test_openapi_schema_controller_tags(openapi_schema: OpenAPI) -> None:
+def test_openapi_schema_controller_tags(openapi_schema: "OpenAPI") -> None:
     assert set(openapi_schema.paths["/controller"].get.tags) == {"handler", "controller"}  # type: ignore
 
 
-def test_openapi_schema_router_tags(openapi_schema: OpenAPI) -> None:
+def test_openapi_schema_router_tags(openapi_schema: "OpenAPI") -> None:
     assert set(openapi_schema.paths["/router/controller"].get.tags) == {"handler", "controller", "router"}  # type: ignore