+++ tests/openapi/utils.py
@@ -3,8 +3,8 @@ from decimal import Decimal
 from enum import Enum
 from typing import Any, Dict, List, Optional, Union
 
-from openapi_schema_pydantic.v3.v3_1_0.example import Example
 from pydantic import conbytes, condecimal, confloat, conint, conlist, conset, constr
+from pydantic_openapi_schema.v3_1_0.example import Example
 
 from starlite import (
     Controller,