+++ tests/plugins/tortoise_orm/test_tortoise_orm_plugin_openapi_specs.py
@@ -1,7 +1,5 @@
 from typing import Callable
 
-import pytest
-
 from starlite import Starlite
 from starlite.plugins.tortoise_orm import TortoiseORMPlugin
 from tests.plugins.tortoise_orm import (
@@ -11,8 +9,7 @@ from tests.plugins.tortoise_orm import (
 )
 
 
-@pytest.mark.asyncio
-async def test_tortoise_orm_plugin_openapi_spec_generation(scaffold_tortoise: Callable) -> None:
+def test_tortoise_orm_plugin_openapi_spec_generation(scaffold_tortoise: Callable) -> None:
     app = Starlite(
         route_handlers=[create_tournament, get_tournament, get_tournaments],
         plugins=[TortoiseORMPlugin()],