+++ tests/route/test_route_map.py
@@ -45,7 +45,7 @@ def route_test_paths(draw: DrawFn) -> List[RouteMapTestCase]:
     components_st = st.lists(st.text(alphabet=ascii_letters, min_size=1, max_size=4), min_size=1, max_size=3)
     path_st = st.builds(build_record, components_st, param_st)
     return cast(
-        List[RouteMapTestCase], draw(st.lists(path_st, min_size=10, max_size=10, unique_by=lambda record: record[1]))
+        "List[RouteMapTestCase]", draw(st.lists(path_st, min_size=10, max_size=10, unique_by=lambda record: record[1]))
     )
 
 
@@ -118,7 +118,8 @@ def test_find_existing_asgi_handlers(test_paths: List[RouteMapTestCase]) -> None
         )
         app.add_node_to_route_map(route)
         asgi_handlers, is_asgi = router.parse_scope_to_route({"path": request_path})
-        assert "GET" in asgi_handlers and isinstance(asgi_handlers["GET"], ExceptionHandlerMiddleware)
+        assert "GET" in asgi_handlers
+        assert isinstance(asgi_handlers["GET"], ExceptionHandlerMiddleware)
         assert is_asgi is False