+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+# Changelog
+
 [0.1.0]
 
 - initial release
@@ -287,3 +289,18 @@
 
 - Allow `Partial` to annotate fields of nested classes @Harry-Lees
 - Add `OpenAPIConfig.use_handler_docstring` param
+
+[1.7.3]
+
+- Fix to routes being allowed under static paths and improvements to path resolution @Dr-Emann
+
+[1.8.0]
+
+- [Breaking] Replace [openapi-pydantic-schema](https://github.com/kuimono/openapi-schema-pydantic)
+  with [pydantic-openapi-schema](https://github.com/starlite-api/pydantic-openapi-schema).
+- [Stoplights Elements](https://stoplight.io/open-source/elements) OpenAPI support @aedify-swi
+
+[1.8.1]
+
+- Add piccolo-orm plugin
+- Fix CacheConfig being broken due to pydantic validation bug